Adding services

+++ erigon.spec
@@ -1,15 +1,20 @@
 Name:           erigon
-vendor:			Ledgerwatch
+Vendor:			Ledgerwatch
 Version:        2022.01.02
 Release:        beta%{?dist}
-Summary:        Ethereum implementation on the efficiency frontier
+Summary:        The Erigon Ethereum Client
 Source0:		%{_sourcedir}/%{name}-%{version}-%{release}.tar.gz
 
 License:        LGPL-3.0
 URL:            https://github.com/ledgerwatch/erigon
 
-BuildRequires:  gcc, gcc-c++, git, gzip, systemd-rpm-macros, rubygem-ronn-ng
-#, golang
+BuildRequires: gcc >= 10
+BuildRequires: gcc-c++ >= 10
+BuildRequires: git
+BuildRequires: gzip
+BuildRequires: systemd-rpm-macros
+BuildRequires: rubygem-ronn-ng
+BuildRequires: golang >= 1.16
 
 %define giturl https://github.com/
 %define repourl %{giturl}%{vendor}/%{name}.git
@@ -20,6 +25,12 @@ BuildRequires:  gcc, gcc-c++, git, gzip, systemd-rpm-macros, rubygem-ronn-ng
 Erigon is an implementation of Ethereum (aka "Ethereum client"), on the efficiency frontier, written in Go.
 
 %prep
+# Build fails with GCC Go, so die unless we can set that alternative:
+GOVER=$(alternatives --list | awk '/^go\s/{print $3}' | grep -v gcc)
+if [ -z "$GOVER" ]; then
+	echo 'Cannot build with GCC-Go! Run "alternatives --config go" and select the official Go binary before rerunning this build!'
+	exit -1
+fi
 # Start as many jobs as we have actual CPU cores
 N_JOBS=$(lscpu -p | grep -v '^#' | cut -d, -f2 | sort -u | wc -l)
 rm -rf %{NVR}
@@ -29,6 +40,7 @@ echo 'Refreshing tags and submodules...'
 git -C %{NVR} pull --recurse-submodules
 git -C %{NVR} fetch --all --tags --recurse-submodules
 git -C %{NVR} submodule update --init --recursive --force
+git -C %{NVR} checkout tags/v%{version} -b v%{version}
 echo 'Building source archive with bundled submodules...'
 # Git strips the NVR because we implicitly change the pwd, but we want to preserve it, thus this hinky business:
 git -C %{NVR} ls-files --recurse-submodules | awk '!/\.git/{print "%{NVR}/"$0}' | tar -T - -czf %{_sourcedir}/%{NVR}.tar.gz
@@ -37,42 +49,58 @@ git -C %{NVR} ls-files --recurse-submodules | awk '!/\.git/{print "%{NVR}/"$0}' 
 # Start as many jobs as we have actual CPU cores
 N_JOBS=$(lscpu -p | grep -v '^#' | cut -d, -f2 | sort -u | wc -l)
 cd %{NVR}
-make -j$N_JOBS erigon rpcdaemon integration sentry txpool downloader hack db-tools
-ln README.md erigon.1.md
-ronn -r --manual 'The Erigon Ethereum Client' --organization Ledgerwatch erigon.1.md
-gzip erigon.1
-rm erigon.1.md
+make -j$N_JOBS %{name} rpcdaemon integration sentry txpool downloader hack db-tools
+ln README.md %{name}.1.md
+ronn -r --manual %{Summary} --organization %{Vendor} %{name}.1.md
+gzip %{name}.1
+rm %{name}.1.md
+# Rename binaries with common names to %{name}_{binary} scheme:
+cd build/bin
+for file in hack integration mdbx_* rpcdaemon downloader sentry txpool; do
+	mv $file %{name}_$file
+done
+cd -
 
 
 %install
-install -m 0755 -D -s %{_builddir}/%{NVR}/build/bin/* -t %{buildroot}/usr/bin
-install -m 0644 -D %{_builddir}/%{NVR}/README.md -t %{buildroot}/usr/share/doc/erigon
-install -m 0644 -D %{_builddir}/%{NVR}/TESTING.md -t %{buildroot}/usr/share/doc/erigon
-install -m 0644 -D %{_builddir}/%{NVR}/COPYING* -t %{buildroot}/usr/share/licenses/erigon
-install -m 0644 -D %{_builddir}/%{NVR}/AUTHORS -t %{buildroot}/usr/share/licenses/erigon
-install -m 0644 -D %{_builddir}/%{NVR}/erigon.1.gz -t %{buildroot}//usr/share/man/man1
+install -m 0755 -D -s %{_builddir}/%{NVR}/build/bin/* -t %{buildroot}%{_bindir}
+install -m 0644 -D %{_builddir}/%{NVR}/README.md -t %{buildroot}%{_datadir}/doc/%{name}
+install -m 0644 -D %{_builddir}/%{NVR}/TESTING.md -t %{buildroot}%{_datadir}/doc/%{name}
+install -m 0644 -D %{_builddir}/%{NVR}/COPYING* -t %{buildroot}%{_datadir}/licenses/%{name}
+install -m 0644 -D %{_builddir}/%{NVR}/AUTHORS -t %{buildroot}%{_datadir}/licenses/%{name}
+install -m 0644 -D %{_builddir}/%{NVR}/%{name}.1.gz -t %{buildroot}%{_mandir}/man1
 
 
 %files
-/usr/bin/erigon
-/usr/bin/hack
-/usr/bin/integration
-/usr/bin/mdbx_chk
-/usr/bin/mdbx_copy
-/usr/bin/mdbx_drop
-/usr/bin/mdbx_dump
-/usr/bin/mdbx_load
-/usr/bin/mdbx_stat
-/usr/bin/pics
-/usr/bin/rpcdaemon
-/usr/bin/rpctest
-/usr/bin/sentry
-/usr/bin/state
-/usr/bin/txpool
-/usr/share/licenses/erigon/AUTHORS
-/usr/share/man/man1/erigon.1.gz
-%license COPYING COPYING.LESSER
-%doc README.md TESTING.md
+%{_bindir}/%{name}
+%{_bindir}/%{name}_hack
+%{_bindir}/%{name}_integration
+%{_bindir}/%{name}_mdbx_chk
+%{_bindir}/%{name}_mdbx_copy
+%{_bindir}/%{name}_mdbx_drop
+%{_bindir}/%{name}_mdbx_dump
+%{_bindir}/%{name}_mdbx_load
+%{_bindir}/%{name}_mdbx_stat
+%{_bindir}/%{name}_rpcdaemon
+%{_bindir}/%{name}_downloader
+%{_bindir}/%{name}_sentry
+%{_bindir}/%{name}_txpool
+%{_datarootdir}/licenses/%{name}/AUTHORS
+%{_mandir}/man1/%{name}.1.gz
+%license %{NVR}/COPYING %{NVR}/COPYING.LESSER
+%doc %{NVR}/README.md %{NVR}/TESTING.md
+
+
+%pre
+if ! getent group %{name} &> /dev/null; then
+	groupadd -r %{name}
+fi
+if getent passwd %{name} &> /dev/null; then
+	mkdir -p %{_sharedstatedir}/%{name}
+	chown -R %{name}:%{name} %{_sharedstatedir}/%{name}
+else
+	useradd -r -m -d %{_sharedstatedir}/%{name} -k /dev/null %{name}
+fi
 
 
 %changelog
